+++ packages/hs-react-ui/src/components/Dropdown/Dropdown.stories.tsx
@@ -10,6 +10,7 @@ import { mdiLeaf } from '@mdi/js';
 import Dropdown, { OptionProps } from './Dropdown';
 import variants from '../../enums/variants';
 import Label from '../Label';
+import { colors } from '../../index';
 
 const design = {
   type: 'figma',
@@ -72,6 +73,7 @@ const teaOptions = [
 ];
 
 storiesOf('Dropdown', module)
+  .addParameters({ component: Dropdown })
   .add(
     'Basic',
     () => {
@@ -80,10 +82,8 @@ storiesOf('Dropdown', module)
         <>
           <Label labelText="City" htmlFor="cities-list">
             <Dropdown
-              // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-              // @ts-ignore - I want to pass undefined to the prop. storybook may not like it but I do.
-              color={color('color', undefined)}
-              elevation={number('elevation', 0, { range: true, min: -5, max: 5, step: 1 })}
+              color={color('color', colors.grayDark)}
+              elevation={number('elevation', 1, { range: true, min: -5, max: 5, step: 1 })}
               multi={boolean('multi', false)}
               name="cities-list"
               placeholder={text('placeholder', 'Choose a city...')}
@@ -95,6 +95,8 @@ storiesOf('Dropdown', module)
               }}
               options={cities}
               variant={select('variant', variants, variants.outline)}
+              optionsVariant={select('optionsVariant', variants, variants.outline)}
+              valueVariant={select('valueVariant', variants, variants.text)}
               values={values}
             />
           </Label>