+++ packages/hs-react-ui/src/components/Progress/Progress.stories.tsx
@@ -9,8 +9,10 @@ import Progress from './Progress';
 //   url: '',
 // };
 
-storiesOf('Progress', module).add(
-  'Basic Progress',
-  () => <Progress />,
-  // { design }, Once we have a design for Progress we can link it here
-);
+storiesOf('Progress', module)
+  .addParameters({ component: Progress })
+  .add(
+    'Basic Progress',
+    () => <Progress />,
+    // { design }, Once we have a design for Progress we can link it here
+  );