+++ packages/hs-react-ui/src/components/RangeSlider/types.ts
@@ -31,12 +31,16 @@ export type RangeSliderProps = {
   StyledSlideRail?: any;
   StyledSelectedRangeRail?: any;
   StyledDomainLabel?: any;
+  StyledMarker?: any;
+  StyledMarkerLabel?: any;
   containerProps?: SubcomponentPropsType;
   dragHandleProps?: SubcomponentPropsType;
   handleLabelProps?: SubcomponentPropsType;
   slideRailProps?: SubcomponentPropsType;
   selectedRangeRailProps?: SubcomponentPropsType;
   domainLabelProps?: SubcomponentPropsType;
+  markerProps?: SubcomponentPropsType;
+  markerLabelProps?: SubcomponentPropsType;
 
   showDomainLabels?: boolean;
   showSelectedRange?: boolean;
@@ -51,4 +55,5 @@ export type RangeSliderProps = {
   max: number;
   values?: number[] | ValueProp[];
   testId?: string;
+  markers?: number[] | ValueProp[];
 };