+++ packages/hs-react-ui/src/components/Tag/Tag.stories.tsx
@@ -1,13 +1,11 @@
 import React from 'react';
-import { action } from '@storybook/addon-actions';
-import { boolean, color, number, select, text } from '@storybook/addon-knobs';
+import { select, text, boolean, number, color } from '@storybook/addon-knobs';
 import { mdiMessage, mdiSend } from '@mdi/js';
 import { storiesOf } from '@storybook/react';
 
-import Button from './Button';
+import Tag from './Tag';
 import colors from '../../enums/colors';
 import variants from '../../enums/variants';
-import FeedbackTypes from '../../enums/feedbackTypes';
 
 const options = {
   none: '',
@@ -17,29 +15,30 @@ const options = {
 
 const design = {
   type: 'figma',
-  url: 'https://www.figma.com/file/3r2G00brulOwr9j7F6JF59/Generic-UI-Style?node-id=83%3A2',
+  url: 'https://www.figma.com/file/3r2G00brulOwr9j7F6JF59/Generic-UI-Style?node-id=0%3A1',
 };
 
-storiesOf('Button', module)
-  .addParameters({ component: Button })
+const testId = `tag-${Math.floor(Math.random() * 100000)}`;
+const containerProps = { 'data-test-id': testId };
+
+storiesOf('Tag', module)
+  .addParameters({ component: Tag })
   .add(
-    'Basic Button',
+    'Basic Tag',
     () => {
       return (
-        <Button
+        <Tag
           variant={select('variant', variants, variants.fill)}
           color={color('color', colors.primaryDark)}
-          onClick={action('button-click')}
-          disabled={boolean('disabled', false)}
-          feedbackType={select('feedbackType', FeedbackTypes, FeedbackTypes.simple)}
           isLoading={boolean('isLoading', false)}
           elevation={number('elevation', 1)}
           isProcessing={boolean('isProcessing', false)}
           iconPrefix={select('iconPrefix', options, options.none)}
           iconSuffix={select('iconSuffix', options, options.none)}
+          containerProps={containerProps}
         >
           {text('children', 'Default text')}
-        </Button>
+        </Tag>
       );
     },
     { design },