+++ packages/hs-react-ui/src/components/TextInput/TextInput.stories.tsx
@@ -1,8 +1,6 @@
 import React, { useState, useCallback } from 'react';
-import styled from 'styled-components';
 import { text, select, boolean, number } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
-import Icon from '@mdi/react';
 import * as IconPaths from '@mdi/js';
 import { storiesOf } from '@storybook/react';
 
@@ -13,6 +11,7 @@ const design = {
   url: 'https://www.figma.com/file/3r2G00brulOwr9j7F6JF59/Generic-UI-Style?node-id=102%3A29',
 };
 storiesOf('TextInput', module)
+  .addParameters({ component: TextInput })
   .add(
     'Basic Text Input',
     () => {
@@ -24,8 +23,6 @@ storiesOf('TextInput', module)
 
       // Setup callbacks to prevent unnecessary rendering
       const onChangeCallback = useCallback(event => {
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
         const newValue = event.target.value;
         setInputValue(newValue);
         action('onChange')(newValue);
@@ -64,7 +61,7 @@ storiesOf('TextInput', module)
 
       return (
         <TextInput
-          ariaLabel={text('ariaLabel', 'textInput', generalGroup)}
+          aria-label={text('ariaLabel', 'textInput', generalGroup)}
           onChange={onChangeCallback}
           debounceInterval={number('debounceInterval', 150, undefined, debounceGroup)}
           debouncedOnChange={onDebounceCallback}
@@ -95,45 +92,6 @@ storiesOf('TextInput', module)
             characterCountGroup,
           )}
           allowTextBeyondMaxLength={boolean('allowTextBeyondMaxLength', false, characterCountGroup)}
-        />
-      );
-    },
-    { design },
-  )
-  .add(
-    'Text Input with all the gadgets',
-    () => {
-      const Input = styled(TextInput.Input)`
-        width: 300px;
-        height: 80px;
-      `;
-      const [inputValue, setInputValue] = useState('');
-      const options = {
-        none: '',
-        ...IconPaths,
-      };
-      const getIconPath = (path: string) => (path ? <Icon size="16px" path={path} /> : undefined);
-      const isMultiline = inputValue.length > 15;
-      const isError =
-        inputValue.length < 5 && inputValue.length > 0 ? 'Short Message Error' : undefined;
-
-      return (
-        <TextInput
-          onChange={event => {
-            const newValue = event.target.value;
-            setInputValue(newValue);
-            action('onChange')(newValue);
-          }}
-          value={inputValue}
-          placeholder={text('placeholder', 'Placeholder')}
-          onClear={() => {
-            setInputValue('');
-            action('onClear')();
-          }}
-          iconPrefix={getIconPath(select('iconPrefix', options, options.mdiComment))}
-          isMultiline={isMultiline}
-          errorMessage={isError}
-          StyledInput={Input}
         />
       );
     },