+++ packages/hs-react-ui/CHANGELOG.md
@@ -1,3 +1,78 @@
+# [1.0.0-v1-release.13](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.12...v1.0.0-v1-release.13) (2020-08-25)
+
+
+### Features
+
+* **dropdown:** add value variants using tags ([1aae340](https://github.com/Headstorm/rasa-ui/commit/1aae340ebe8918986c55136ac5b207b84351f22e))
+
+# [1.0.0-v1-release.12](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.11...v1.0.0-v1-release.12) (2020-08-24)
+
+
+### Features
+
+* **docspage props table:** add props table to docspage ([09b124f](https://github.com/Headstorm/rasa-ui/commit/09b124f428fe0971dd7a6dc999c494102f06d112))
+
+# [1.0.0-v1-release.11](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.10...v1.0.0-v1-release.11) (2020-08-21)
+
+
+### Bug Fixes
+
+* **textinput:** spread the rest of the props onto the Input element ([856f6cb](https://github.com/Headstorm/rasa-ui/commit/856f6cb5804b1c265c790b8649a94ee1e2550fae))
+
+# [1.0.0-v1-release.10](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.9...v1.0.0-v1-release.10) (2020-08-20)
+
+
+### Features
+
+* **card:** add interaction feedback ([946f907](https://github.com/Headstorm/rasa-ui/commit/946f9071f6837cb4b8ae4103a2e814e302b8d490))
+
+# [1.0.0-v1-release.9](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.8...v1.0.0-v1-release.9) (2020-08-20)
+
+
+### Features
+
+* **button, interactionfeedback:** add InteractionFeedback wrapper around Button ([26cf13d](https://github.com/Headstorm/rasa-ui/commit/26cf13da518827176d696842ea79e6e67a1d2447))
+
+# [1.0.0-v1-release.8](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.7...v1.0.0-v1-release.8) (2020-08-19)
+
+
+### Features
+
+* **rangeslider:** add markers API ([519f096](https://github.com/Headstorm/rasa-ui/commit/519f096712f300ac76ef1675a1dee1ab82b4f07d))
+
+# [1.0.0-v1-release.7](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.6...v1.0.0-v1-release.7) (2020-08-19)
+
+
+### Bug Fixes
+
+* **textinput:** fix text input charCounter not appearing at the top w/ no error message ([4f31ea2](https://github.com/Headstorm/rasa-ui/commit/4f31ea2ed9b460ae406aee59138a7a36f2b77a5e))
+
+
+### Features
+
+* **dropdown:** add variants to dropdown ([0cc22b3](https://github.com/Headstorm/rasa-ui/commit/0cc22b3978ed9fb51e086adb64c75b44348578fa))
+
+# [1.0.0-v1-release.6](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.5...v1.0.0-v1-release.6) (2020-08-19)
+
+
+### Features
+
+* **rangeslider:** add functionality to support clicking the RangeSlider's rail to set the value ([beeb2e6](https://github.com/Headstorm/rasa-ui/commit/beeb2e625bd37b3f7d23dea10378adc695ebe3ba))
+
+# [1.0.0-v1-release.5](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.4...v1.0.0-v1-release.5) (2020-08-19)
+
+
+### Features
+
+* **tag:** add new Tag component ([2fa81a4](https://github.com/Headstorm/rasa-ui/commit/2fa81a42e8f7ad9bdff609382e208e5020ab63b8))
+
+# [1.0.0-v1-release.4](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.3...v1.0.0-v1-release.4) (2020-08-19)
+
+
+### Bug Fixes
+
+* **table:** remove row height of 100% ([1d10f24](https://github.com/Headstorm/rasa-ui/commit/1d10f24294873fd6b9145efd198efe9cce527d1c))
+
 # [1.0.0-v1-release.3](https://github.com/Headstorm/rasa-ui/compare/v1.0.0-v1-release.2...v1.0.0-v1-release.3) (2020-08-18)